extract transaction slice
